src/main/java/ru/maliutin/firstrestapp/services/PeopleService.java
package ru.maliutin.firstrestapp.services;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import ru.maliutin.firstrestapp.models.Person;
import ru.maliutin.firstrestapp.repositories.PeopleRepository;
import ru.maliutin.firstrestapp.util.PersonNotFoundException;

import java.util.List;
import java.util.Optional;

@Service
@Transactional(readOnly = true)
public class PeopleService {
    private final PeopleRepository peopleRepository;

    @Autowired
    public PeopleService(PeopleRepository peopleRepository) {
        this.peopleRepository = peopleRepository;
    }

    public List<Person> findAll() {
        return peopleRepository.findAll();
    }

    /*
        В данном методе реализована логика возврата объекта Person по его id,
        если такого человека не найдено - генерируется исключение PersonNotFoundException
     */
    public Person findOne(int id) {
        Optional<Person> person = peopleRepository.findById(id);
        return person.orElseThrow(PersonNotFoundException::new);
    }
}
